requirement.txt
Flask==2.2.2
pymongo==4.3.3
Requests==2.28.1
pycryptodome==3.17
numpy==1.23.5
opencv-python==4.6.0.66